+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes to be released
 
 ## Fix
 - Issue where sound would not play when using mod #4
+- Issue where gui entry for keybinding would not update
 
 ## Add
 - Turkish Translation